+++ src/hassette/core/classes/app/app.py
@@ -6,17 +6,18 @@ from typing import Any, ClassVar, Generic
 from anyio import to_thread
 
 from hassette.config.app_manifest import AppManifest
-from hassette.core.bus import Bus
 from hassette.core.classes.resource import Resource
 from hassette.core.enums import ResourceRole
 from hassette.core.events import Event
-from hassette.core.scheduler import Scheduler
 
 from .app_config import AppConfig, AppConfigT
 from .utils import validate_app
 
 if typing.TYPE_CHECKING:
+    from hassette.core.api import Api
+    from hassette.core.bus import Bus
     from hassette.core.core import Hassette
+    from hassette.core.scheduler import Scheduler
 
 LOGGER = getLogger(__name__)
 
@@ -61,10 +62,13 @@ class App(Generic[AppConfigT], Resource):
     logger: logging.Logger
     """Logger for the instance."""
 
-    scheduler: Scheduler
+    api: "Api"
+    """API instance for interacting with Home Assistant."""
+
+    scheduler: "Scheduler"
     """Scheduler instance for scheduled tasks owned by this app."""
 
-    bus: Bus
+    bus: "Bus"
     """Event bus instance for event handlers owned by this app."""
 
     def __init_subclass__(cls, **kwargs):
@@ -79,14 +83,7 @@ class App(Generic[AppConfigT], Resource):
             cls._import_exception = e
             LOGGER.exception("Failed to initialize subclass %s", cls.__name__)
 
-    def __init__(
-        self,
-        hassette: "Hassette",
-        app_config: AppConfigT,
-        index: int = 0,
-        *args,
-        **kwargs,
-    ):
+    def __init__(self, hassette: "Hassette", app_config: AppConfigT, index: int = 0):
         """Initialize the App instance. This will generally not be called directly.
 
         Args:
@@ -95,35 +92,21 @@ class App(Generic[AppConfigT], Resource):
             index (int): Index of the app instance, used when multiple instances of the same app are run.
 
         """
+        from hassette.core import Api, Bus, Scheduler
+
+        super().__init__(hassette=hassette, unique_name_prefix=f"{self.class_name}.{app_config.instance_name}")
+
         self.app_config = app_config
         self.index = index
 
-        super().__init__(hassette=hassette, *args, **kwargs)  # noqa: B026
-
-        logger_name = f"hassette.{type(self).__name__}.{self.app_config.instance_name}"
-        self.logger = getLogger(logger_name)
-
-        if self.instance_name == "":
-            # this shouldn't happen, as AppHandler will set a default if not provided
-            raise ValueError("App instance_name cannot be empty.")
-
-        self.bus = Bus(self.hassette, owner=self.unique_name)
-        self.scheduler = Scheduler(self.hassette, owner=self.unique_name)
-
-    @property
-    def unique_name(self) -> str:
-        """Unique identifier for the instance."""
-        return f"{self.class_name}-{self.instance_name}-{self.unique_id}"
+        self.bus = Bus(self.hassette, owner=f"{self.unique_name}.bus", task_bucket=self.task_bucket)
+        self.scheduler = Scheduler(self.hassette, owner=f"{self.unique_name}.scheduler", task_bucket=self.task_bucket)
+        self.api = Api(self.hassette, unique_name_prefix=f"{self.unique_name}.api", task_bucket=self.task_bucket)
 
     @property
     def instance_name(self) -> str:
         """Name for the instance of the app. Used for logging and ownership of resources."""
         return self.app_config.instance_name
-
-    @property
-    def api(self):
-        """API instance for interacting with Home Assistant."""
-        return self.hassette.api
 
     async def send_event(self, event_name: str, event: Event[Any]) -> None:
         """Send an event to the event bus."""