+++ src/hassette/core/classes/base.py
@@ -1,7 +1,7 @@
 import copy
 import typing
 import uuid
-from logging import getLogger
+from logging import Logger, getLogger
 
 from typing_extensions import deprecated
 
@@ -11,40 +11,25 @@ if typing.TYPE_CHECKING:
     from hassette.core.core import Hassette
 
 
-class _HassetteBase:  # pyright: ignore[reportUnusedClass]
-    class_name: typing.ClassVar[str]
-    """Name of the class, set on subclassing."""
+class _LoggerMixin:
+    """Mixin to provide logging capabilities to classes."""
 
-    role: typing.ClassVar[ResourceRole] = ResourceRole.BASE
-    """Role of the resource, e.g. 'App', 'Service', etc."""
+    unique_id: str
+    """Unique identifier for the instance."""
 
-    def __init_subclass__(cls) -> None:
-        cls.class_name = cls.__name__
+    logger: Logger
+    """Logger for the instance."""
 
-    def __init__(self, hassette: "Hassette", *args, **kwargs) -> None:
-        """
-        Initialize the class with a reference to the Hassette instance.
+    unique_name: str
+    """Unique name for the instance."""
 
-        Args:
-            hassette (Hassette): The Hassette instance this resource belongs to.
-            *args: Additional positional arguments.
-            **kwargs: Additional keyword arguments.
-        """
-
+    def __init__(self, unique_name_prefix: str | None = None) -> None:
         self.unique_id = uuid.uuid4().hex
-        """Unique identifier for the instance."""
+        self.unique_name = f"{unique_name_prefix or type(self).__name__}.{self.unique_id[:8]}"
+        self.logger = getLogger(f"hassette.{self.unique_name}")
 
-        self.hassette = hassette
-        """Reference to the Hassette instance."""
-
-        self.logger = getLogger(f"{type(self).__module__}.{self.class_name}")
-
-        self.logger.debug("Creating instance of '%s'", self.class_name)
-
-    @property
-    def unique_name(self) -> str:
-        """Unique identifier for the instance."""
-        return f"{self.class_name}-{self.unique_id}"
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__} unique_name={self.unique_name}>"
 
     def set_logger_to_level(self, level: typing.Literal["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]) -> None:
         """Configure a logger to log at the specified level independently of its parent."""
@@ -71,3 +56,32 @@ class _HassetteBase:  # pyright: ignore[reportUnusedClass]
     def set_logger_to_debug(self) -> None:
         """Configure a logger to log at DEBUG level independently of its parent."""
         self.set_logger_to_level("DEBUG")
+
+
+class _HassetteBase(_LoggerMixin):  # pyright: ignore[reportUnusedClass]
+    class_name: typing.ClassVar[str]
+    """Name of the class, set on subclassing."""
+
+    role: typing.ClassVar[ResourceRole] = ResourceRole.BASE
+    """Role of the resource, e.g. 'App', 'Service', etc."""
+
+    hassette: "Hassette"
+    """Reference to the Hassette instance."""
+
+    def __init_subclass__(cls) -> None:
+        cls.class_name = cls.__name__
+
+    def __init__(self, hassette: "Hassette", unique_name_prefix: str | None = None) -> None:
+        """
+        Initialize the class with a reference to the Hassette instance.
+
+        Args:
+            hassette (Hassette): The Hassette instance this resource belongs to.
+            unique_name_prefix (str | None): Optional prefix for the unique name. If None, the class name is used.
+        """
+        super().__init__(unique_name_prefix=unique_name_prefix)
+        self.hassette = hassette
+        self.logger.debug("Creating instance of '%s'", self.class_name)
+
+    def __repr__(self) -> str:
+        return f"<{self.class_name} unique_name={self.unique_name}>"